가맹검품은 수기마감대상 수정 -> NativeQuery, 확인품등록 카카오톡메세지 -> 가맹점의대한 메세지전송은 제외, 가맹점 frMultiscreen 사용여부 api 추가, 가맹점 고객등록 전화번호중복체크 수정

+++ src/main/java/com/broadwave/toppos/User/ReuqestMoney/Requset/RequestDetail/RequestDetailDtos/user/RequestDetailCloseListDto.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 /**
  * @author Minkyu
  * Date : 2022-01-18
@@ -16,7 +18,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class RequestDetailCloseListDto {
-    private Long fdId; // 고유ID값
+    private BigInteger fdId; // 고유ID값
 
     private String frYyyymmdd; // 접수일자
     private String bcName; // 고객명
